Fix content script item extraction and API validation
- Fix bbox type from List[int] to List[float] (getBoundingClientRect returns floats)
- Reduce size threshold for link elements (10x10 vs 50x20) to capture HN-style links
- Add debug logging for content script initialization and filtering stats
- Add service worker message logging for debugging

+++ interestlens/backend/models/requests.py
@@ -15,7 +15,7 @@ class PageItem(BaseModel):
     href: Optional[str] = None
     text: str
     snippet: str = ""
-    bbox: List[int] = []  # [x, y, width, height]
+    bbox: List[float] = []  # [x, y, width, height] - floats from getBoundingClientRect
     thumbnail_base64: Optional[str] = None
